Move StateT instances to contrib.scalaz
