+++ connector/src/test/scala/quasar/fs/FileSystemFixture.scala
@@ -21,6 +21,7 @@ import quasar.Predef._
 import quasar.Data
 import quasar.DataArbitrary._
 import quasar.contrib.pathy._, PathArbitrary._
+import quasar.contrib.scalaz.stateT.stateTCatchable
 import quasar.fp._
 import quasar.fp.free.{Interpreter, SpecializedInterpreter}
 import quasar.fs.SandboxedPathy._
