+++ connector/src/test/scala/quasar/fs/ManageFileSpec.scala
@@ -19,6 +19,7 @@ package quasar.fs
 import quasar.Predef._
 import quasar.Data
 import quasar.contrib.scalaz.eitherT._
+import quasar.contrib.scalaz.stateT._
 import quasar.fp._, ski._
 
 import scala.Predef.$conforms
