+++ connector/src/test/scala/quasar/fs/QueryFileSpec.scala
@@ -21,6 +21,7 @@ import quasar.{Data, DataArbitrary}
 import quasar.common.PhaseResults
 import quasar.contrib.pathy._
 import quasar.contrib.scalaz.eitherT._
+import quasar.contrib.scalaz.stateT._
 import quasar.fp._
 import quasar.frontend.logicalplan.{LogicalPlan, LogicalPlanR}
 import quasar.scalacheck._
