+++ connector/src/test/scala/quasar/fs/WriteFileSpec.scala
@@ -21,6 +21,7 @@ import quasar.Predef._
 import quasar.{Data, DataArbitrary}
 import quasar.contrib.pathy._
 import quasar.contrib.scalaz.eitherT._
+import quasar.contrib.scalaz.stateT._
 import quasar.fp._
 
 import org.specs2.specification.core._
