+++ foundation/src/main/scala/quasar/contrib/scalaz/package.scala
@@ -17,46 +17,10 @@
 package quasar.contrib
 
 import _root_.scalaz._, \&/._
-import _root_.scalaz.syntax.monad._
 
 package object scalaz {
   def -\&/[A, B](a: A): These[A, B] = This(a)
   def \&/-[A, B](b: B): These[A, B] = That(b)
-
-  implicit def stateTMonadError[F[_], E, S](implicit F: MonadError[F, E])
-      : MonadError[StateT[F, S, ?], E] =
-    new MonadError[StateT[F, S, ?], E] {
-      def point[A](a: => A) = a.point[F].liftM[StateT[?[_], S, ?]]
-
-      def bind[A, B]
-        (fa: StateT[F, S, A])
-        (f: A => StateT[F, S, B]) =
-        StateT[F, S, B](s =>
-          F.bind(fa.run(s))(p => f(p._2).run(p._1)))
-
-      def handleError[A]
-        (fa: StateT[F, S, A])
-        (f: E => StateT[F, S, A]) =
-        StateT[F, S, A](s => F.handleError(fa.run(s))(f(_).run(s)))
-
-      def raiseError[A](e: E) =
-        StateT[F, S, A](_ => F.raiseError[(S, A)](e))
-    }
-
-  implicit def stateTMonadTell[F[_], W, S](implicit F: MonadTell[F, W])
-      : MonadTell[StateT[F, S, ?], W] =
-    new MonadTell[StateT[F, S, ?], W] {
-      def point[A](a: => A) = a.point[F].liftM[StateT[?[_], S, ?]]
-
-      def bind[A, B]
-        (fa: StateT[F, S, A])
-        (f: A => StateT[F, S, B]) =
-        StateT[F, S, B](s =>
-          F.bind(fa.run(s))(p => f(p._2).run(p._1)))
-
-      def writer[A](w: W, v: A): StateT[F, S, A] =
-        F.writer(w, v).liftM[StateT[?[_], S, ?]]
-    }
 
   implicit def toMonadTell_Ops[F[_], W, A](fa: F[A])(implicit F: MonadTell_[F, W]) =
     new MonadTell_Ops[F, W, A](fa)
